chore(web): deduplicate API error helpers
