+++ packages/franken-web/src/lib/dashboard-api.ts
@@ -1,3 +1,5 @@
+import { extractResponseErrorMessage, toError } from './http-error';
+
 export interface DashboardSkill {
   name: string;
   enabled: boolean;
@@ -140,28 +142,9 @@ export class DashboardApiClient {
 
 }
 
-function toError(error: unknown): Error {
-  return error instanceof Error ? error : new Error(String(error));
-}
 
 async function createResponseError(res: Response): Promise<Error> {
   const message = await extractResponseErrorMessage(res);
   return new Error(message ?? `HTTP ${res.status}`);
 }
 
-async function extractResponseErrorMessage(res: Response): Promise<string | undefined> {
-  try {
-    const body = await res.json() as unknown;
-    if (!body || typeof body !== 'object') return undefined;
-    const error = (body as { error?: unknown }).error;
-    if (typeof error === 'string' && error.trim()) return error;
-    if (error && typeof error === 'object') {
-      const message = (error as { message?: unknown }).message;
-      if (typeof message === 'string' && message.trim()) return message;
-    }
-  } catch {
-    return undefined;
-  }
-
-  return undefined;
-}
